requirements.txt
langgraph
langchain-google-genai
langchain-core
langchain-community
python-dotenv
google-generativeai
mysql-connector-python